Main page 요소들 컴퍼넌트화
