+++ src/pages/Mun/Main/FeedBox/FeedBox.scss
@@ -1,0 +1,125 @@
+.feedBox {
+  margin: 50px 30px 0px 0px;
+  border: 1px solid #e6e6e6;
+  width: 500px;
+  background-color: #ffffff;
+
+  &:nth-last-of-type(1) {
+    margin-bottom: 30px;
+  }
+
+  .feedTop {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    .feedTopLeft {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      font-size: 13px;
+
+      img {
+        margin: 10px;
+        border-radius: 70%;
+        overflow: hidden;
+        height: 30px;
+        width: 30px;
+      }
+    }
+
+    .feedTopRight {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+
+      img {
+        margin: 10px;
+        height: 25px;
+      }
+    }
+  }
+
+  .feedMiddleImg {
+    width: 498px;
+    height: 498px;
+
+    img {
+      width: 100%;
+      height: 100%;
+    }
+  }
+
+  .bottomMenu {
+    display: flex;
+    justify-content: space-between;
+    height: 35px;
+    padding: 0 10px;
+
+    img {
+      height: 20px;
+      margin-bottom: 15px;
+    }
+
+    .bottomMenuLeft {
+      img {
+        margin-right: 15px;
+      }
+    }
+  }
+
+  .like {
+    font-size: 13px;
+    padding: 5px 10px;
+    display: flex;
+    align-items: center;
+
+    img {
+      width: 20px;
+      height: 20px;
+      border-radius: 70%;
+      overflow: hidden;
+      margin-right: 5px;
+    }
+  }
+
+  .commentContainer {
+    .commentMore {
+      font-size: 12px;
+      font-weight: bold;
+      color: #9e9e9e;
+      padding: 0 10px;
+    }
+  }
+
+  .commentTime {
+    font-size: 8px;
+    color: #9e9e9e;
+    padding: 5px 12px 0px 12px;
+  }
+
+  .commentInputBox {
+    width: 100%;
+    height: 40px;
+    border-top: 1px solid #e6e6e6;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 10px;
+    margin-top: 10px;
+
+    .commentInput {
+      border: 0;
+      outline: 0;
+      width: 85%;
+    }
+
+    .commentButton {
+      border: 0;
+      outline: 0;
+      background-color: transparent;
+      color: #c4e1fb;
+      cursor: pointer;
+    }
+  }
+}
