+++ src/pages/Mun/Main/Main.scss
@@ -1,254 +1,31 @@
-/* --------------------네비게이션 바-------------------- */
-
 .mainContainer {
   padding-top: 45px;
   font-size: 50px;
   display: flex;
   justify-content: center;
   background-color: #fafafa;
-  .feedContainer {
-    .feedBox {
-      margin: 50px 30px 0px 0px;
-      border: 1px solid #e6e6e6;
-      width: 500px;
-      background-color: #ffffff;
-      &:nth-last-of-type(1) {
-        margin-bottom: 30px;
-      }
-      .feedTop {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        .feedTopLeft {
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-          font-size: 13px;
-          img {
-            margin: 10px;
-            border-radius: 70%;
-            overflow: hidden;
-            height: 30px;
-            width: 30px;
-          }
-        }
-        .feedTopRight {
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-          img {
-            margin: 10px;
-            height: 25px;
-          }
-        }
-      }
-      .feedMiddleImg {
-        width: 498px;
-        height: 498px;
-        img {
-          width: 100%;
-          height: 100%;
-        }
-      }
-      .bottomMenu {
-        display: flex;
-        justify-content: space-between;
-        height: 35px;
-        padding: 0 10px;
-        img {
-          height: 20px;
-          margin-bottom: 15px;
-        }
-        .bottomMenuLeft {
-          img {
-            margin-right: 15px;
-          }
-        }
-      }
-      .like {
-        font-size: 13px;
-        padding: 5px 10px;
-        display: flex;
-        align-items: center;
-        img {
-          width: 20px;
-          height: 20px;
-          border-radius: 70%;
-          overflow: hidden;
-          margin-right: 5px;
-        }
-      }
-      .commentContainer {
-        .commentMore {
-          font-size: 12px;
-          font-weight: bold;
-          color: #9e9e9e;
-          padding: 0 10px;
-        }
-      }
-      .commentTime {
-        font-size: 8px;
-        color: #9e9e9e;
-        padding: 5px 12px 0px 12px;
-      }
-      .commentInputBox {
-        width: 100%;
-        height: 40px;
-        border-top: 1px solid #e6e6e6;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        padding: 0 10px;
-        margin-top: 10px;
-        .commentInput {
-          border: 0;
-          outline: 0;
-          width: 85%;
-        }
-        .commentButton {
-          border: 0;
-          outline: 0;
-          background-color: transparent;
-          color: #c4e1fb;
-          cursor: pointer;
-        }
-      }
-    }
-  }
+
   .asideFeed {
     width: 300px;
     margin-top: 50px;
-    .meBox {
-      display: flex;
-      align-items: center;
-      height: 60px;
-      img {
-        width: 45px;
-        height: 45px;
-        border-radius: 70%;
-        margin: 10px;
-      }
-      .meNameBox {
-        .meEngName {
-          font-size: 15px;
-          padding: 2px;
-          font-weight: bold;
-        }
-        .meName {
-          font-size: 15px;
-          padding: 2px;
-          color: #9e9e9e;
-        }
-      }
-    }
-    .storyBox {
-      border: 1.5px solid #e6e6e6;
-      background-color: #ffffff;
-      border-radius: 4px;
-      height: 220px;
-      overflow: hidden;
-      .asideBoxTop {
-        display: flex;
-        justify-content: space-between;
-        font-size: 13px;
-        padding: 10px 15px;
-        font-weight: bold;
-      }
-      .asideUserSpace {
-        display: flex;
-        align-items: center;
-        height: 50px;
-        img {
-          width: 40px;
-          height: 40px;
-          border-radius: 70%;
-          margin: 0px 15px;
-        }
-        .asideUser {
-          font-size: 13px;
-          .userName {
-            font-weight: bold;
-            padding: 1px;
-            padding-right: 4px;
-          }
-          .asideUserTime {
-            color: #9e9e9e;
-            padding: 1px;
-          }
-        }
-      }
-    }
-    .recommendBox {
-      border: 1.5px solid #e6e6e6;
-      background-color: #ffffff;
-      border-radius: 4px;
-      height: 200px;
-      overflow: hidden;
-      margin-top: 11px;
-      .asideBoxTop {
-        display: flex;
-        justify-content: space-between;
-        font-size: 13px;
-        padding: 10px 15px;
-        font-weight: bold;
-      }
-      .asideUserSpace {
-        display: flex;
-        align-items: center;
-        height: 50px;
-        img {
-          width: 40px;
-          height: 40px;
-          border-radius: 70%;
-          margin: 0px 15px;
-        }
-        .asideUser {
-          font-size: 13px;
-          .userName {
-            font-weight: bold;
-            padding: 1px;
-            padding-right: 4px;
-          }
-          .asideUserFriend {
-            display: inline-block;
-            color: #9e9e9e;
-            padding: 1px;
-            width: 167px;
-            white-space: nowrap;
-            overflow: hidden;
-            text-overflow: ellipsis;
-          }
-        }
-        .followButton {
-          border: 0;
-          outline: 0;
-          background-color: transparent;
-          color: #0095f6;
-        }
-      }
-    }
-    .noticeBox {
-      width: 80%;
-      color: #dcdcdc;
-      font-size: 15px;
-      .noticeInfo {
-        margin: 13px 0px 15px 0px;
-      }
-    }
   }
 }
+
 .commentClose {
   height: 8px;
   width: 8px;
   margin-right: 6px;
   opacity: 0.4;
+
   &:hover {
     cursor: pointer;
   }
 }
+
 .grayFont {
   color: #9e9e9e;
 }
+
 .userName {
   font-weight: bold;
   padding: 1px;
