+++ src/pages/Mun/Main/NoticeBox/NoticeBox.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import './NoticeBox.scss';
+
+const NoticeBox = () => {
+  return (
+    <footer className="noticeBox">
+      <div className="noticeInfo">
+        Instagram 정보·지원·홍보 센터·API·채용
+        정보·개인정보처리방침·약관·디렉터리·프로필·해시태그·언어
+      </div>
+      <div>ⓒ 2019 INSTAGRAM</div>
+    </footer>
+  );
+};
+
+export default NoticeBox;
