+++ src/pages/Mun/Main/NoticeBox/NoticeBox.scss
@@ -1,0 +1,9 @@
+.noticeBox {
+  width: 80%;
+  color: #dcdcdc;
+  font-size: 15px;
+
+  .noticeInfo {
+    margin: 13px 0px 15px 0px;
+  }
+}
