+++ src/pages/Mun/Main/ProfileBox/ProfileBox.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import './ProfileBox.scss';
+const ProfileBox = () => {
+  return (
+    <div className="profileBox">
+      <div>
+        <img src="./images/Mun/meimg.jpg" alt="tmp" />
+      </div>
+      <div className="myNameBox">
+        <div className="myEngName">nyam_nyam2</div>
+        <div className="myKoName">냠냠이</div>
+      </div>
+    </div>
+  );
+};
+
+export default ProfileBox;
