+++ src/pages/Mun/Main/ProfileBox/ProfileBox.scss
@@ -1,0 +1,26 @@
+.profileBox {
+  display: flex;
+  align-items: center;
+  height: 60px;
+
+  img {
+    width: 45px;
+    height: 45px;
+    border-radius: 70%;
+    margin: 10px;
+  }
+
+  .myNameBox {
+    .myEngName {
+      font-size: 15px;
+      padding: 2px;
+      font-weight: bold;
+    }
+
+    .myKoName {
+      font-size: 15px;
+      padding: 2px;
+      color: #9e9e9e;
+    }
+  }
+}
