+++ src/pages/Mun/Main/StoryBox/StoryBox.scss
@@ -1,0 +1,43 @@
+.storyBox {
+  border: 1.5px solid #e6e6e6;
+  background-color: #ffffff;
+  border-radius: 4px;
+  height: 220px;
+  overflow: hidden;
+
+  .asideBoxTop {
+    display: flex;
+    justify-content: space-between;
+    font-size: 13px;
+    padding: 10px 15px;
+    font-weight: bold;
+  }
+
+  .asideUserSpace {
+    display: flex;
+    align-items: center;
+    height: 50px;
+
+    img {
+      width: 40px;
+      height: 40px;
+      border-radius: 70%;
+      margin: 0px 15px;
+    }
+
+    .asideUser {
+      font-size: 13px;
+
+      .userName {
+        font-weight: bold;
+        padding: 1px;
+        padding-right: 4px;
+      }
+
+      .asideUserTime {
+        color: #9e9e9e;
+        padding: 1px;
+      }
+    }
+  }
+}
